Deprecate `use-add-on-display-cost`
